Interest: Enable interest days to be passed via constructor

+++ midas-api/src/main/java/de/nihas101/midas/api/interest/InterestCalculationRow.java
@@ -11,7 +11,7 @@ public interface InterestCalculationRow {
 
     Transaction balanceAtEndOfMonth();
 
-    Integer interestDaysCount();
+    BigDecimal interestDaysCount();
 
     BigDecimal interestAmount();
