+++ midas-core/src/main/java/de/nihas101/midas/core/interest/row/DivisorRow.java
@@ -41,7 +41,7 @@ public class DivisorRow implements InterestCalculationRow {
     }
 
     @Override
-    public Integer interestDaysCount() {
+    public BigDecimal interestDaysCount() {
         return null;
     }
