+++ midas-core/src/main/java/de/nihas101/midas/core/interest/row/InterestRow.java
@@ -45,7 +45,7 @@ public class InterestRow implements InterestCalculationRow {
     }
 
     @Override
-    public Integer interestDaysCount() {
+    public BigDecimal interestDaysCount() {
         return null;
     }
