+++ midas-core/src/main/java/de/nihas101/midas/core/interest/row/InterestSumRow.java
@@ -41,7 +41,7 @@ public class InterestSumRow implements InterestCalculationRow {
     }
 
     @Override
-    public Integer interestDaysCount() {
+    public BigDecimal interestDaysCount() {
         return null;
     }
